+++ src/App.js
@@ -1,4 +1,4 @@
-import "./App.css";
+import "./css/App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
 import { Routes, Route } from "react-router-dom";
 import Home from "./routes/Home";
