+++ src/index.js
@@ -4,7 +4,7 @@ import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import axios from "axios";
-import "./index.css";
+import "./css/index.css";
 
 axios.defaults.baseURL = process.env.REACT_APP_API_URL;
